// web/src/hooks/queries/gql-generated-uniswap.ts
export type Maybe<T> = T | null;
// ...
export type Bundle = {
  __typename?: "Bundle";
  ethPriceUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
};
// ...
export type Burn = {
  __typename?: "Burn";
  amount: Scalars["BigInt"]["output"];
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD?: Maybe<Scalars["BigDecimal"]["output"]>;
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  origin: Scalars["Bytes"]["output"];
  owner?: Maybe<Scalars["Bytes"]["output"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"]["output"];
  tickUpper: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Collect = {
  __typename?: "Collect";
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD?: Maybe<Scalars["BigDecimal"]["output"]>;
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  owner?: Maybe<Scalars["Bytes"]["output"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"]["output"];
  tickUpper: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type Factory = {
  __typename?: "Factory";
  id: Scalars["ID"]["output"];
  owner: Scalars["ID"]["output"];
  poolCount: Scalars["BigInt"]["output"];
  totalFeesETH: Scalars["BigDecimal"]["output"];
  totalFeesUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedETH: Scalars["BigDecimal"]["output"];
  totalValueLockedETHUntracked: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  totalVolumeETH: Scalars["BigDecimal"]["output"];
  totalVolumeUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type Flash = {
  __typename?: "Flash";
  amount0: Scalars["BigDecimal"]["output"];
  amount0Paid: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amount1Paid: Scalars["BigDecimal"]["output"];
  amountUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  pool: Pool;
  recipient: Scalars["Bytes"]["output"];
  sender: Scalars["Bytes"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type Mint = {
  __typename?: "Mint";
  amount: Scalars["BigInt"]["output"];
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD?: Maybe<Scalars["BigDecimal"]["output"]>;
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  origin: Scalars["Bytes"]["output"];
  owner: Scalars["Bytes"]["output"];
  pool: Pool;
  sender?: Maybe<Scalars["Bytes"]["output"]>;
  tickLower: Scalars["BigInt"]["output"];
  tickUpper: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export enum OrderDirection {
  Asc = "asc",
  Desc = "desc",
}

export type Pool = {
  __typename?: "Pool";
  burns: Array<Burn>;
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  collectedFeesUSD: Scalars["BigDecimal"]["output"];
  collects: Array<Collect>;
  createdAtBlockNumber: Scalars["BigInt"]["output"];
  createdAtTimestamp: Scalars["BigInt"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feeTier: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  liquidityProviderCount: Scalars["BigInt"]["output"];
  mints: Array<Mint>;
  observationIndex: Scalars["BigInt"]["output"];
  poolDayData: Array<PoolDayData>;
  poolHourData: Array<PoolHourData>;
  sqrtPrice: Scalars["BigInt"]["output"];
  swaps: Array<Swap>;
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  ticks: Array<Tick>;
  token0: Token;
  token0Price: Scalars["BigDecimal"]["output"];
  token1: Token;
  token1Price: Scalars["BigDecimal"]["output"];
  totalValueLockedETH: Scalars["BigDecimal"]["output"];
  totalValueLockedToken0: Scalars["BigDecimal"]["output"];
  totalValueLockedToken1: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type PoolDayData = {
  __typename?: "PoolDayData";
  close: Scalars["BigDecimal"]["output"];
  date: Scalars["Int"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  high: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  low: Scalars["BigDecimal"]["output"];
  open: Scalars["BigDecimal"]["output"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"]["output"];
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  token0Price: Scalars["BigDecimal"]["output"];
  token1Price: Scalars["BigDecimal"]["output"];
  tvlUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type PoolHourData = {
  __typename?: "PoolHourData";
  close: Scalars["BigDecimal"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  high: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  low: Scalars["BigDecimal"]["output"];
  open: Scalars["BigDecimal"]["output"];
  periodStartUnix: Scalars["Int"]["output"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"]["output"];
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  token0Price: Scalars["BigDecimal"]["output"];
  token1Price: Scalars["BigDecimal"]["output"];
  tvlUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type Position = {
  __typename?: "Position";
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  depositedToken0: Scalars["BigDecimal"]["output"];
  depositedToken1: Scalars["BigDecimal"]["output"];
  feeGrowthInside0LastX128: Scalars["BigInt"]["output"];
  feeGrowthInside1LastX128: Scalars["BigInt"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  owner: Scalars["Bytes"]["output"];
  pool: Pool;
  tickLower: Tick;
  tickUpper: Tick;
  token0: Token;
  token1: Token;
  transaction: Transaction;
  withdrawnToken0: Scalars["BigDecimal"]["output"];
  withdrawnToken1: Scalars["BigDecimal"]["output"];
};
// ...
export type Query = {
  __typename?: "Query";
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  bundle?: Maybe<Bundle>;
  bundles: Array<Bundle>;
  burn?: Maybe<Burn>;
  burns: Array<Burn>;
  collect?: Maybe<Collect>;
  collects: Array<Collect>;
  factories: Array<Factory>;
  factory?: Maybe<Factory>;
  flash?: Maybe<Flash>;
  flashes: Array<Flash>;
  mint?: Maybe<Mint>;
  mints: Array<Mint>;
  pool?: Maybe<Pool>;
  poolDayData?: Maybe<PoolDayData>;
  poolDayDatas: Array<PoolDayData>;
  poolHourData?: Maybe<PoolHourData>;
  poolHourDatas: Array<PoolHourData>;
  pools: Array<Pool>;
  position?: Maybe<Position>;
  positionSnapshot?: Maybe<PositionSnapshot>;
  positionSnapshots: Array<PositionSnapshot>;
  positions: Array<Position>;
  swap?: Maybe<Swap>;
  swaps: Array<Swap>;
  tick?: Maybe<Tick>;
  tickDayData?: Maybe<TickDayData>;
  tickDayDatas: Array<TickDayData>;
  tickHourData?: Maybe<TickHourData>;
  tickHourDatas: Array<TickHourData>;
  ticks: Array<Tick>;
  token?: Maybe<Token>;
  tokenDayData?: Maybe<TokenDayData>;
  tokenDayDatas: Array<TokenDayData>;
  tokenHourData?: Maybe<TokenHourData>;
  tokenHourDatas: Array<TokenHourData>;
  tokens: Array<Token>;
  transaction?: Maybe<Transaction>;
  transactions: Array<Transaction>;
  uniswapDayData?: Maybe<UniswapDayData>;
  uniswapDayDatas: Array<UniswapDayData>;
};
// ...
export type Swap = {
  __typename?: "Swap";
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  origin: Scalars["Bytes"]["output"];
  pool: Pool;
  recipient: Scalars["Bytes"]["output"];
  sender: Scalars["Bytes"]["output"];
  sqrtPriceX96: Scalars["BigInt"]["output"];
  tick: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Token = {
  __typename?: "Token";
  decimals: Scalars["BigInt"]["output"];
  derivedETH: Scalars["BigDecimal"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  poolCount: Scalars["BigInt"]["output"];
  symbol: Scalars["String"]["output"];
  tokenDayData: Array<TokenDayData>;
  totalSupply: Scalars["BigInt"]["output"];
  totalValueLocked: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volume: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
  whitelistPools: Array<Pool>;
};
// ...
export type _Meta_ = {
  __typename?: "_Meta_";
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars["String"]["output"];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars["Boolean"]["output"];
};

export enum _SubgraphErrorPolicy_ {
  /** Data will be returned even if the subgraph has indexing errors */
  Allow = "allow",
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  Deny = "deny",
}
// ...
export type GetPoolHourDatasQuery = {
  __typename?: "Query";
  poolHourDatas: Array<{
    __typename?: "PoolHourData";
    token0Price: string;
    token1Price: string;
    periodStartUnix: number;
    sqrtPrice: string;
    liquidity: string;
    pool: {
      __typename?: "Pool";
      id: string;
      liquidity: string;
      token0: { __typename?: "Token"; id: string; name: string };
      token1: { __typename?: "Token"; id: string; name: string };
    };
  }>;
};
// ...
export const GetPoolHourDatasDocument = gql`
    query GetPoolHourDatas($skip: Int = 0, $first: Int, $where: PoolHourData_filter, $orderBy: PoolHourData_orderBy, $orderDirection: OrderDirection, $block: Block_height, $subgraphError: _SubgraphErrorPolicy_! = deny) {
  poolHourDatas(
    skip: $skip
    first: $first
    orderBy: $orderBy
    orderDirection: $orderDirection
    where: $where
    block: $block
    subgraphError: $subgraphError
  ) {
    token0Price
    token1Price
    periodStartUnix
    sqrtPrice
    liquidity
    pool {
      id
      liquidity
      token0 {
        id
        name
      }
      token1 {
        id
        name
      }
    }
  }
}
    `;